Add max-age caching support
This is a naive implementation while we are looking for a new HTTP request library with native caching support.

It's not possible to interrupt an axios request and respond from cache using interceptors. So this PR drops almost all interceptors in favor of koa-style middlewares. That will also make our future migration away from axios easier.