Add production info to workspace creation call
We will no longer support changing the `production` flag of a workspace. This
means that we need to be able to create the workspace with `production` both set
to `true` or `false`, instead of always setting it to false.

According to the new guidelines we only allow links on workspaces that have
`production` set to `false` and we only allow workspaces with `production` set
to `true` to be promoted. This is going to be implemented in toolbelt, but first
we need to allow it to create both kinds of workspace.