Increase free sockets lifetime to 60sec
Router instances are less volatile so we can use a longer
timeout.

Also, when keep-alived the sockets are sent a TCP keepalive
packet every 1 sec (the default for keepAliveMsec opts), so
we will still avoid using "bad" connections since as soon
as the keep alive fails we will throw them out.