Updates log client initialization
This update includes a timeout to handle clients initializing indefinitely, assuming some initial logs might be lost on application startup while the log client still loading. Now we can better handle this scenario since we have more control over the startup process of the log client.
We've also updated error handling for this step, avoiding throwing errors on log client startup so the app is still available and able to use console.log fallback - ensuring the app is able to run even if the logger is not working properly.